requirements.txt
fastapi
uvicorn
python-multipart
jinja2

presidio-analyzer
presidio-anonymizer

faker
python-docx
pandas

transformers
torch
seqeval
spacy

snowflake-snowpark-python[pandas]

streamlit
python-dotenv
cryptography
